Started working on a ChartAppCreator class that pulls up an interactive window that lets you choose the functions you want to graph

+++ src/main/java/JavaPSSFIles/FunctionPlot.java
@@ -7,10 +7,10 @@ import java.io.IOException;
 import java.util.ArrayList;
 public class FunctionPlot {
     //This will produce the plot points for the first 25 points of an exponential function
-    public ArrayList<double[]> expFunc(int initialX){
+    public ArrayList<double[]> expFunc(double initialX){
         ArrayList<double[]> points = new ArrayList<>();
         double y;
-        for(int i = initialX; i < 25; i++){
+        for(double i = initialX; i < 25; i++){
             y = Math.exp(i);
             initialX = i;
             points.add(new double[]{initialX,y});
